Tidied up some definitions and added isomorphism target
I've moved conversion-related stuff to Conversion.agda. In there is an
IsoProof record, one of which I am trying to construct for Nat. Right now
I am having trouble with the InvNat proof.

Basically invProofA and invProofRepA are functions that, given a value of
the relevant type, return a proof of the relevant inverse property.